+++ clone_service.cpp
@@ -22,7 +22,7 @@ using Association = std::tuple<std::string, std::string, std::string>;
 using Value = std::variant<bool, uint8_t, int16_t, uint16_t, int32_t, uint32_t,
                            int64_t, uint64_t, double, std::string,
                            std::vector<std::string>, std::vector<Association>,
-                           std::vector<uint8_t>,
+                           std::vector<uint8_t>, std::vector<double>,
                            std::vector<int16_t>, std::vector<uint16_t>,
                            std::vector<int32_t>, std::vector<uint32_t>,
                            std::vector<int64_t>, std::vector<uint64_t>>;
@@ -178,6 +178,10 @@ std::string variantToString(const Value &variantVar)
     else if (std::holds_alternative<std::vector<uint64_t>>(variantVar))
     {
         value = declareArray<std::vector<uint64_t>>(variantVar, 'T');
+    }
+    else if (std::holds_alternative<std::vector<double>>(variantVar))
+    {
+        value = declareArray<std::vector<double>>(variantVar, 'D');
     }
     else if(std::holds_alternative<std::vector<Association>>(variantVar))
     {